refactor to chained slave model
