+++ src/main/java/jtp/cc1/ascsp1/MonsterGameServlet.java
@@ -9,9 +9,6 @@ interface Scene {
   //public String draw(Game g, Game[] savedGames);
   //public Scene whereToNext(Game g, String input);
   //public void doGet(HttpServletRequest req, HttpServletResponse resp);
-  default String json(String screen, String method, String other) {
-    return "{ \"screen\": \"" + screen + "\", \"method\": \"" + method + "\", \"other\": \"" + other + "\" }";
-  }
   
 }
